Embeded lab results into patient record.
